src/third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/event-timing/dblclick-expected.txt

This is a testharness.js-based test.
FAIL Event Timing dblclick. assert_greater_than: The entry's processingStart should be greater than startTime. expected a number greater than 832.0800000001327 but got 832.0800000001327
Harness: the test ran to completion.
